pyproject.toml
[project]
name = "brownfield-cartographer"
version = "0.1.0"
description = "Multi-agent codebase intelligence system for rapid FDE onboarding"
readme = "README.md"
authors = [
    {name = "Tsegay", email = "tsegayassefa27@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    # Core
    "typing-extensions>=4.15.0",
    "pydantic>=2.12.5",
    "pydantic-core>=2.41.5",
    
    # Static Analysis - tree-sitter
    "tree-sitter>=0.21.3",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-sql>=0.1.0",
    "tree-sitter-yaml>=0.5.0",
    "tree-sitter-typescript>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    
    # Graph Analysis
    "networkx>=3.2.1",
    
    # Git Analysis
    "gitpython>=3.1.46",
    
    # Data Lineage (partial for Phase 1)
    "sqlglot>=26.11.0",
    
    # Utilities
    "pyyaml>=6.0.3",
    "click>=8.1.7",
    "pathspec>=0.12.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
cartographer = "src.cli:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "ARG"]
ignore = ["B904"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov=src --cov-report=term-missing"
testpaths = ["tests"]
python_files = "test_*.py"
